Added Controllers and Stuff

+++ server/src/com/opentill/models/BaseModels/BaseModel.java
@@ -23,12 +23,12 @@ public abstract class BaseModel {
 	public int updated;
 	
 	@Version
-    private Long version;
+    private int version;
 	
-	public Long getVersion() {
+	public int getVersion() {
 		return version;
 	}
-	public void setVersion(Long version) {
+	public void setVersion(int version) {
 		this.version = version;
 	}
 	public boolean deleted;
